requirements.txt
fastapi
uvicorn
torch
opencv-python
ultralytics
python-multipart
Pillow